+++ internal/e2e_harness/production/schema_evolution_e2e_test.go
@@ -234,7 +234,7 @@ func TestSchemaEvolutionAddedColumn(t *testing.T) {
 		map[string]string{"name": "VARCHAR", "value": "DOUBLE", "label": "VARCHAR"})
 	forbidParquetCols(t, "base (v1)", baseCols, "score")
 	requireParquetCols(t, "delta (v2)", describeParquetCols(ctx, t, env, deltaKey),
-		map[string]string{"score": "INTEGER", "label": "VARCHAR"})
+		map[string]string{"score": "BIGINT", "label": "VARCHAR"})
 
 	full := env.AssertQueryMatches(ctx, Query{Schema: simple, Limit: 20})
 	assertUsesDuckDB(t, full)
@@ -283,7 +283,7 @@ func TestSchemaEvolutionRemovedColumn(t *testing.T) {
 	seedGeneration(ctx, t, env, simple, 5, scoreProfile)
 	baseKey := runInitBase(ctx, t, env, simple)
 	requireParquetCols(t, "base (v1)", describeParquetCols(ctx, t, env, baseKey),
-		map[string]string{"score": "INTEGER", "label": "VARCHAR"})
+		map[string]string{"score": "BIGINT", "label": "VARCHAR"})
 
 	if err := env.EvolveSchema(ctx, v2); err != nil {
 		t.Fatalf("evolve schema to v2: %v", err)
@@ -327,14 +327,14 @@ func TestSchemaEvolutionRemovedColumn(t *testing.T) {
 
 // TestSchemaEvolutionChangedType covers #189 scenario 3: `score` keeps its
 // attributeID but its valueType changes integer→numeric, so the v1 base
-// parquet stores INTEGER where the v2 delta stores DOUBLE. Contract:
+// parquet stores BIGINT where the v2 delta stores DOUBLE. Contract:
 // predictable widening — the union resolves to DOUBLE, matching the oracle's
 // numeric-family float64 normalization; filters and sorts see one coherent
 // numeric domain across generations.
 //
 // Red on current main with SILENTLY WRONG DATA, not a loud failure: without
 // union_by_name DuckDB coerces every file to the first file's schema, so the
-// delta's DOUBLE values are cast to the base's INTEGER — fractional scores
+// delta's DOUBLE values are cast to the base's BIGINT — fractional scores
 // are corrupted in place and only the oracle catches it (attr mismatches,
 // totals identical). That failure mode is the strongest reason the fix must
 // widen the union rather than pin first-file-wins.
@@ -355,7 +355,7 @@ func TestSchemaEvolutionChangedType(t *testing.T) {
 		})))
 
 	requireParquetCols(t, "base (v1 integer)", describeParquetCols(ctx, t, env, baseKey),
-		map[string]string{"score": "INTEGER"})
+		map[string]string{"score": "BIGINT"})
 	requireParquetCols(t, "delta (v2 numeric)", describeParquetCols(ctx, t, env, deltaKey),
 		map[string]string{"score": "DOUBLE"})
 
@@ -366,7 +366,7 @@ func TestSchemaEvolutionChangedType(t *testing.T) {
 	}
 
 	// One numeric domain across generations: the threshold catches v1
-	// INTEGER rows (20,30,40) and every v2 DOUBLE row (50.5..110.5).
+	// BIGINT rows (20,30,40) and every v2 DOUBLE row (50.5..110.5).
 	filtered := env.AssertQueryMatches(ctx, Query{
 		Schema:  simple,
 		Filters: []Filter{{Attr: "score", Op: "gte", Value: "20"}},
@@ -408,7 +408,7 @@ func TestSchemaEvolutionMixedGenerations(t *testing.T) {
 	requireParquetCols(t, "base (v1)", baseCols, map[string]string{"old_col": "VARCHAR"})
 	forbidParquetCols(t, "base (v1)", baseCols, "new_col")
 	deltaCols := describeParquetCols(ctx, t, env, deltaKey)
-	requireParquetCols(t, "delta (v2)", deltaCols, map[string]string{"new_col": "INTEGER"})
+	requireParquetCols(t, "delta (v2)", deltaCols, map[string]string{"new_col": "BIGINT"})
 	forbidParquetCols(t, "delta (v2)", deltaCols, "old_col")
 
 	full := env.AssertQueryMatches(ctx, Query{Schema: simple, Limit: 20})
@@ -461,10 +461,10 @@ func TestSchemaEvolutionRenamedColumn(t *testing.T) {
 		})))
 
 	baseCols := describeParquetCols(ctx, t, env, baseKey)
-	requireParquetCols(t, "base (v1)", baseCols, map[string]string{"score": "INTEGER"})
+	requireParquetCols(t, "base (v1)", baseCols, map[string]string{"score": "BIGINT"})
 	forbidParquetCols(t, "base (v1)", baseCols, "points")
 	deltaCols := describeParquetCols(ctx, t, env, deltaKey)
-	requireParquetCols(t, "delta (v2)", deltaCols, map[string]string{"points": "INTEGER"})
+	requireParquetCols(t, "delta (v2)", deltaCols, map[string]string{"points": "BIGINT"})
 	forbidParquetCols(t, "delta (v2)", deltaCols, "score")
 
 	full := env.AssertQueryMatches(ctx, Query{Schema: simple, Limit: 20})
